refactor: organize `tts` namespace
